Added default logging
/Api/RestHost has LogError() Method, which can be overriden;  However,
the /Api/Host class does not override; nor, makes it possible for user
of the class to log anything.
This update introduces a LogErr Deligate, which enables users of the
class to define their own logging mechanism.  Additionally, if no
logging delegate is supplied, we will default to errors to IIS W3SVC log
file so that integration issues can be identified easier.